backend/requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
python-dotenv==1.0.0
bcrypt==5.0.0
pydantic[email]==2.12.5
firebase-admin==7.3.0
pymongo==4.6.0
motor==3.3.2
